feat(schema, sql): BigInt を 64 bit 列 1 本に (主キー・group・合計・上位 k 件も) + SQL BIGINT (段階 E)
schema の ColumnType::BigInt (`.bigint(col)`) を engine の Number64 の列 1 本にする。 値は大小の順を保つ
符号化 (`v ^ 2^63`) の u64 で置くので、 範囲・並び・上位 k 件がそのまま効く。 値域は BIGINT_MIN..=BIGINT_MAX
(= i64::MIN..=i64::MAX - 1、 i64::MAX だけは engine の空の印と重なる)。

- 主キーにできる (upsert の照合 / sync の PK bind #141 は query_by_id64)、 group の列・合計の列・上位 k 件の購読
- 条件の値を u64 に (Predicate::Eq / Range / In)、 where_gt / where_ge / where_lt / where_le / where_range の値は
  Into<Value> (i64 を渡せる、 u32 / i32 の呼び出しはそのまま)、 Number 列では値域で切る。 内部の Cmp は範囲に寄せた
- find は query_by_id64 / pull_in_by_id64 (engine に追加) / get64、 並びは engine の値で
- 64 bit の集計: Query::sum_i128 / min_i64 / max_i64 (Number / BigInt)。 u32 の集計は BigInt で BadValue
- LiveCounts の合計は i128。 BigInt の合計は engine の Agg::summed (合計に足した件数、 engine に追加) で元の値に
  戻す。 summed だけが動いた (値 0 ↔ 値なし) 報告は (件数, 合計) が同じなら渡さない
- engine: Agg::summed、 合計する値は 「値 + 1 (0 = 値なし)」 で持ち値の無い根は足さない。 count_eids_par は 64 bit
  列を u32 の配列で読まない
- SQL: BIGINT 列 (i64、 INTEGER は u32 のまま)。 負の数値リテラルは符号ごと読む (i64::MIN)、 INSERT / UPDATE は
  書く前に全部の値を確かめる (途中の列で失敗して半端な row を残さない — INTEGER の範囲外でも残っていた)

テスト: schema tests/bigint.rs 5 本 (値と条件の oracle、 購読の oracle、 reopen / add_column、 主キー・64 bit の
集計・group / 合計 / 上位 k 件、 BigInt で group / 負の数を含む合計 / 上位 k 件の購読の oracle)、
tests/bigint_sync.rs 2 本 (値が届く・書き合っても書いた値そのもの、 BigInt の主キーの束ね)、 sql の BIGINT 1 本。
変異 8 本 (符号化 / 戻し / 合計の戻し / 主キーの照合 / 並び / sync の照合 / SQL の戻し / SQL の事前確認) で落ちる。
workspace 1221 passed

+++ crates/enchudb-oplog/src/oplog.rs
@@ -1864,7 +1864,7 @@ mod tests {
         let wal = OpLog::create(&p, HEADER_SIZE + 256).unwrap();
         let mut hit_full = false;
         for i in 0..10_000u64 {
-            match wal.append(Op::Tie { eid: i, himo_id: 0, value: i as u64 }) {
+            match wal.append(Op::Tie { eid: i, himo_id: 0, value: i }) {
                 Ok(_) => {}
                 Err(e) => {
                     assert_eq!(e.kind(), io::ErrorKind::OutOfMemory, "full は OutOfMemory で返る");
@@ -2098,7 +2098,7 @@ mod tests {
                     w.append(Op::Tie {
                         eid: t * 10000 + i,
                         himo_id: 0,
-                        value: i as u64,
+                        value: i,
                     }).unwrap();
                 }
             }));
@@ -2129,7 +2129,7 @@ mod tests {
                     w.append(Op::Tie {
                         eid: t * 10_000 + i,
                         himo_id: 0,
-                        value: i as u64,
+                        value: i,
                     }).unwrap();
                 }
             }));
